refactor: don't colour command names, add hidden commands
As we were colouring the command names, when commander tried to look the name provided up it would
fail as the name commander was provided would include ansi codes. The help looks a little less
jazzy now but we can live with that.

Also add hidden commands for the hidden commands as commander will now fail as it doesn't know
about these